+++ src/app/search/search-ingredient-folder.tsx
@@ -13,7 +13,11 @@ export function SearchIngredientFolder({ category, folded, onChange }: SearchIng
     return (
         <span className={styles['search-ingredient-category__ingredient']} key={ingredientId}>
             <input type="checkbox" name={ingredientId} id={ingredientId} className="d-none" onChange={onChange} />
-            <label htmlFor={ingredientId}>{folded ? 'Rozwiń' : 'Zwiń'}</label>
+            <label htmlFor={ingredientId}>
+                <div className="d-flex justify-content-center align-items-center w-100 h-100">
+                    {folded ? 'Rozwiń' : 'Zwiń'}
+                </div>
+            </label>
         </span>
     );
 }